% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_bias.r
\name{aggregate_to_cluster}
\alias{aggregate_to_cluster}
\title{aggregate_to_cluster}
\usage{
aggregate_to_cluster(model)
}
\arguments{
\item{model}{a meta3 model}
}
\description{
Aggregate effect sizes to cluster
}
\details{
multiple effect sizes per cluster are aggregated using a fixed effects meta-analysis
}
