% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{as_word}
\alias{as_word}
\title{as_word}
\usage{
as_word(x = NULL, sentence = F, hyphenate = T)
}
\arguments{
\item{x}{integer to convert to a word}

\item{sentence}{if TRUE capitalise the first letter}

\item{hyphenate}{should hyphens be included?}
}
\description{
Convert numbers to words
}
