% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate.r
\name{coef.meta_list}
\alias{coef.meta_list}
\title{coef.meta_list}
\usage{
\method{coef}{meta_list}(object, ...)
}
\arguments{
\item{object}{meta_list object}

\item{...}{unused}
}
\description{
coef.meta_list
}
