% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{dummy_matrix}
\alias{dummy_matrix}
\title{dummy_matrix}
\usage{
dummy_matrix(x, levels = NULL, pattern = ",")
}
\arguments{
\item{x}{a character vector}

\item{levels}{a vector. Provides the level (and order) of colnames.}

\item{pattern}{pattern provided to str_split}
}
\description{
creates a predictor matrix when cells can contain multiple tags
}
