% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{forest_height}
\alias{forest_height}
\title{forest_height}
\usage{
forest_height(meta3L_plot, slope = 0.12, intercept = 0.52)
}
\arguments{
\item{meta3L_plot}{a moderated meta3L plot}

\item{slope}{the numeric value to multiple number of rows by}

\item{intercept}{the numeric constant}
}
\description{
Use this function to resize forest plots in rmarkdown.
}
