% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{forest_plot_data}
\alias{forest_plot_data}
\title{forest_plot_data}
\usage{
forest_plot_data(
  model,
  moderators,
  transf,
  baseline_name,
  author,
  year,
  envir = parent.frame()
)
}
\arguments{
\item{model}{meta3 or meta_list}

\item{moderators}{a string of moderators}

\item{transf}{function to transform results}

\item{baseline_name}{string to rename baseline model}

\item{author}{string, name of author column}

\item{year}{string, name of year column}

\item{envir}{environment where source dataset is located}
}
\description{
Create data ready for the creation of a forest plot
}
