% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{funnel_plot}
\alias{funnel_plot}
\title{funnel_plot}
\usage{
funnel_plot(
  model,
  xlab = "Estimate",
  ylab = "Standard error",
  font = "serif",
  aggregate = FALSE,
  trimfill = FALSE,
  alpha = 1,
  size = 1.5,
  CI_linetype = "dashed",
  CI_size = 0.7,
  pool_linetype = "dotted",
  pool_size = 0.5,
  density = FALSE,
  ...
)
}
\arguments{
\item{model}{an object belonging to the class 'meta_list'. These objects are created by the function 'meta3_moderation'.}

\item{xlab}{a character string. Label for the x-axis.}

\item{ylab}{a character string. Label for the y-axis.}

\item{font}{a character string. Set's font family. Defaults to times new roman ('serif')}

\item{aggregate}{should data be aggregated to cluster level?}

\item{trimfill}{should trimfill be performed?}

\item{alpha}{a scalar. Sets alpha transparency.}

\item{size}{a scalar. The ggplot2 size value for the points}

\item{CI_linetype}{ggplot2 linetype for confidence intervals}

\item{CI_size}{ggplot2 size value for confidence intervals}

\item{pool_linetype}{ggplot2 linetype for pooled estimate}

\item{pool_size}{ggplot2 size value for pooled estimate}

\item{density}{a bool. If True, plots densit of points.}

\item{...}{extra arguments passed to rma.uni for the purpose of trimfill.}
}
\description{
This function is used for plotting funnel plots
}
\details{
produces a funnel plot using ggplot2. If aggregate is set to TRUE, data is aggregated using fixed-effects meta-analysis to each cluster level (the pooled effect is not updated). If trimfill is set to TRUE, extra observations will be created as needed to achieve symmetry using metafor::trimfill. If trimfill results in new studies, the pooled effect will be updated based on the results of that meta-analysis—which is not multi-level. Note that funnel plots are often difficult to interpret, especially for multi-level meta-analysis. Note that the assumptions of trim and fill are violated if there are dependencies within the data (e.g., if data is clustered and you used multi-level meta-analysis to address this clustering). Rows with missing y or v are removed.
}
