% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{get_kn}
\alias{get_kn}
\title{get_kn}
\usage{
get_kn(model, param_names)
}
\arguments{
\item{model}{model to extract kn from}

\item{param_names}{the names of slopes}
}
\description{
get numbers of studies and effect sizes
}
