% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_val}
\alias{get_val}
\title{get_val}
\usage{
get_val(model, val, moderator = NULL, round = 2, transf = function(x) x)
}
\arguments{
\item{model}{the meta3 or meta_list model to extract}

\item{val}{character vector of values to extract}

\item{moderator}{character string of moderator to extract from. Defaults to NULL}

\item{round}{2 number of digits}

\item{transf}{function by which to transform relevant estimates (only applies to estimate and estimate95)}
}
\description{
Return values
}
