% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{meta3_OK}
\alias{meta3_OK}
\title{meta3_OK}
\usage{
meta3_OK(m)
}
\description{
Check whether meta3 status is OK
}
\details{
This function looks up the OpenMx status code. If 0 or 1 returns TRUE, else returns FALSE
}
