% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{meta3L_OK}
\alias{meta3L_OK}
\title{meta3L_OK}
\usage{
meta3L_OK(m)
}
\arguments{
\item{m}{meta3L model}
}
\description{
Check whether meta3L status is OK
}
\details{
This function looks up the OpenMx status code. If 0 or 1 returns TRUE, else returns FALSE
}
