% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{meta_matrix}
\alias{meta_matrix}
\title{meta_matrix}
\usage{
meta_matrix(formula, data, intercept = FALSE, warn = TRUE)
}
\arguments{
\item{formula}{formula yi + vi + cluster ~ predictor1 + predictor_n}

\item{data}{data object}

\item{intercept}{bool, TRUE includes an intercept}

\item{warn}{bool. If TRUE, you will be warned about variables with no variance}
}
\description{
Creates a matrix for use in the meta-analysis
}
