% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{model.matrix2}
\alias{model.matrix2}
\title{model.matrix2}
\usage{
model.matrix2(
  object,
  data = environment,
  contrasts.args = NULL,
  xlev = NULL,
  ...
)
}
\arguments{
\item{object}{a formula}

\item{data}{data.frame object}

\item{contrasts.args}{argument passed to model.matrix}

\item{xlev}{argument passed to model.matrix}

\item{...}{other arguments passed to model.matrix}
}
\description{
model.matrix but does not assign a prefix to dummy variables
}
