% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate.r
\name{moderate}
\alias{moderate}
\title{moderate}
\usage{
moderate(m, ..., moderators = NULL, debug = FALSE)
}
\arguments{
\item{m}{a meta3 object}

\item{...}{a named list of moderators}

\item{moderators}{You can feed a named character vector for consistent moderators across models}

\item{debug}{If true, ... contents are returned}
}
\description{
Create a list of meta3 objects moderated by predictors
}
