% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-and-methods.R
\name{plot.meta_list}
\alias{plot.meta_list}
\title{meta_list plot method}
\usage{
\method{plot}{meta_list}(x, ...)
}
\arguments{
\item{x}{model to print}

\item{...}{additional arguments passed to ninjaForest.}
}
\description{
meta_list plot method
}
