% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate.r
\name{print.KIN_summary}
\alias{print.KIN_summary}
\title{print.KIN_summary}
\usage{
\method{print}{KIN_summary}(x, ...)
}
\arguments{
\item{x}{meta_list object}

\item{...}{additional arguments passed to format_nicely}
}
\description{
print.KIN_summary
}
