% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate.r
\name{print.meta_list}
\alias{print.meta_list}
\title{meta_list methods}
\usage{
\method{print}{meta_list}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{additional arguments. Not currently used.}
}
\description{
meta_list methods
}
