% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{print.mlm_summary}
\alias{print.mlm_summary}
\title{print.mlm_summary}
\usage{
\method{print}{mlm_summary}(x, ...)
}
\arguments{
\item{x}{mlm object}

\item{...}{additional arguments passed to print}
}
\description{
Print function for mlm_summary
}
