% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_n}
\alias{report_n}
\title{report_n}
\usage{
report_n(x, rmarkdown = FALSE)
}
\arguments{
\item{x}{meta3 model}

\item{rmarkdown}{if TRUE, rmarkdown code is produced which can be edited}
}
\description{
Get number of studies and effect sizes of a baseline model
}
