% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate.r
\name{round_p}
\alias{round_p}
\title{round_p}
\usage{
round_p(
  p,
  n = 3,
  stars = c(),
  leading.zero = T,
  apa_threshold = 0.001,
  simplify = 0.1
)
}
\arguments{
\item{p}{numeric}

\item{n}{number of digits}

\item{stars}{numeric vector for thresholds after which stars will be added. e.g. 0.05}

\item{leading.zero}{If FALSE the leading zero is removed}

\item{apa_threshold}{numeric. The number after which results are summarised with less than symbols}

\item{simplify}{numeric. Defines the threshold above which a digit is dropped}
}
\description{
Rounds p using APA formatting
}
