% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate.r
\name{summary.meta_list}
\alias{summary.meta_list}
\title{summary.meta_list}
\usage{
\method{summary}{meta_list}(object, ...)
}
\arguments{
\item{object}{meta_list object}

\item{...}{additional arguments passed to format_nicely}
}
\description{
summary.meta_list
}
