% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{summary.metalm}
\alias{summary.metalm}
\title{summary.metalm}
\usage{
\method{summary}{metalm}(object, ...)
}
\arguments{
\item{object}{metalm object}

\item{...}{additional arguments passed to model_slopes}
}
\description{
Summary function for metalm
}
