% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_bias.r
\name{trim_and_fill}
\alias{trim_and_fill}
\title{trim_and_fill}
\usage{
trim_and_fill(model, aggregate = TRUE, ...)
}
\arguments{
\item{model}{a model of class meta3L}

\item{aggregate}{should data be aggregated first?}

\item{...}{additional arguments passed to metafor::rma.uni}
}
\value{
rma.uni.trimfill
}
\description{
Perform trim and fill for meta3L objects.
}
\details{
uses data aggregated by cluster.
}
