% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{try_even_harder}
\alias{try_even_harder}
\title{try_even_harder}
\usage{
try_even_harder(model, extraTries = 25)
}
\arguments{
\item{model}{the model.}

\item{extraTries}{the number of extra tries}

\item{autofixtau2}{Logical. Whether automatically fixes elements of tau2 with NA of standard errors. It only works for objects of class tssem1REM, class meta, and class osmasem.}
}
\description{
reruns models with problems
}
