% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.R
\name{apa_table.metaKIN_table}
\alias{apa_table.metaKIN_table}
\title{apa_table.metaKIN_table}
\usage{
apa_table.metaKIN_table(
  x,
  caption = NULL,
  note = NULL,
  added_stub_head = NULL,
  col_spanners = NULL,
  midrules = NULL,
  placement = "tbp",
  landscape = FALSE,
  font_size = NULL,
  escape = FALSE,
  span_text_columns = TRUE,
  format.args = NULL,
  ...
)
}
\description{
apa_table.metaKIN_table
}
