% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{forest_plot}
\alias{forest_plot}
\title{forest_plot}
\usage{
forest_plot(
  model,
  ...,
  xlab = "Effect size",
  effect_label = NULL,
  transf = NULL,
  baseline_name = "Pooled estimate",
  xlim = NULL,
  facet_by = NULL,
  vline = 0,
  author = NULL,
  year = NULL,
  moderator.shape = 23,
  moderator.size = 3,
  summary.shape = 23,
  summary.size = 4,
  moderator_diamond = FALSE,
  font = "serif",
  envir = parent.frame(),
  return_data = FALSE
)
}
\arguments{
\item{model}{a meta_list object}

\item{...}{moderators to plot character strings only}

\item{xlab}{a string.}

\item{effect_label}{a string. If provided relabels "effect size" in facetting.}

\item{transf}{a function. If supplied effect sizes are transformed by this function}

\item{baseline_name}{a string. The label for the baseline model}

\item{xlim}{passed to scale_x_continous}

\item{facet_by}{a colname. Facets effect sizes by a supplied variable.}

\item{vline}{a scalar. Dictates the x-intercept (the dashed line).}

\item{author}{the name of the author column}

\item{year}{the name of the year column}

\item{moderator.shape}{a scalar ggplot2 geom_point shape value}

\item{moderator.size}{a scalar. ggplot2 geom_point size value}

\item{summary.shape}{a scalar ggplot2 geom_point shape value}

\item{summary.size}{a scalar ggplot2 geom_point size value}

\item{moderator_diamond}{a bool. If true, diamond is created for moderators}

\item{font}{A string. The name of a font family. Defaults to serif}

\item{envir}{the calling environment to retrieve model data by name}

\item{return_data}{if TRUE, plotting data is returned and plotting does not occur}
}
\description{
Produce a forest plot for meta_list objects
}
