% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{mlm}
\alias{mlm}
\title{mlm}
\usage{
mlm(m, formula, model.name = NULL, .envir = parent.frame())
}
\arguments{
\item{m}{meta3 object}

\item{formula}{formula passed to model.matrix}

\item{model.name}{String. Name your model (optional)}

\item{.envir}{the environment to run in}
}
\description{
Meta regression for meta3 objects
}
