% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssem_variables.R
\name{report_tssem2}
\alias{report_tssem2}
\title{report_tssem2}
\usage{
report_tssem2(x, pattern = NULL)
}
\arguments{
\item{x}{a wls object}

\item{pattern}{A glue pattern}
}
\description{
Report the results of tssem
}
